Fix for is_android_raw in detecting *.apk files
The previous version of 'is_android_raw' is not able to detect some *.apk files in which the string 'PK' is not in the first block of 7 bytes. 
So, to correctly detect all *.apk files we need also to check some immutable resources which are absolutely present in their packages, such as 'AndroidManifest.xml' and 'META-INF/MANIFEST.MF'.