create 'androguard apkid' command for quite APK identification
This first tries to do quick binary XML parsing to just get the
values that are needed.  It will fallback to full androguard
parsing, which is slow, if it can't find the versionName value or
versionName is set to a Android String Resource (e.g. an integer
hex value that starts with @).