Fixing Session module
this resolves #435 but adds some new behaviour to the Session
module:

If an APK is added to the Session, analyzed_vms contains now 1+n
entries, where n is the number of DEX files inside the APK.
Before there would be only one entry for the APK but no entries for
the DEX files.

All functions relying on analyzed_vms will return the first
occurence of the Analysis inside this array, thus usually will
return the Digest of the APK and not the digest of the DEX file
associated with the Analysis object.