Add new xrefs: new_instance, const_class and xref_read/write to MethodAnalysis
To avoid iterating over all fields in order to find the ones a method
can read/write from/to, the Fields are now available through a
MethodAnalysis object. Additionally, the _new_instance and _cons_class
references are added to MethodAnalysis and ClassAnalysis for the same
reason as well as to avoid using xref_to to a non-existent method.