Set custom directory for JADX source files
JADX by default places the decompiled sources into a "sources" directory
inside the directory passed by the "-d" flag. The "-ds" flag can be used
to specify a custom directory for the source files. By setting it to the
tmpfolder, JADX  will place the source files where Androguard expects
them to be.