APKv1 signatures with a cert chain are parsed like apksigner
Microsoft signs their APKs with a X.509 certificate chain of trust, so
there are actually three certificates included. apksigner only cares
about the certificate that actually signs the manifest and ignores the
certificates that just sign other certificates.

https://apkpure.com/microsoft-edge-preview/com.microsoft.emmx/download
https://gitlab.com/fdroid/fdroidserver/-/issues/1128
https://gitlab.com/fdroid/fdroidserver/-/merge_requests/1466

X.509 certificates and JAR signatures are a combination of machine-
generated and plain data with trivial human input, so are not
copyrightable.  So I included signature files directly.