Convert `var_decl` into a static method
This commit is part of a series of changes that convert functions in `androguard.decompiler.dast` into instance or static methods.

With the previous version of the code it was difficult to alter the functionality of the `JSONWriter` class.

For some methods in `JSONWriter`, expanding the behaviour would require re-implementing entire code blocks or using monkey patching.

Example:

In the original version of the code the `dast` module contains a function called `method_invocation`. It is used in another function called `visit_expr`. `visit_expr` is called by `JSONWriter` in `get_cond` and `visit_switch_node` instance methods.

If a developer wanted to add extra logic for cases when `JSONWriter` detects a method invocation (e.g. to dynamically log callers), the procedure would be very complicated:

1. Use some form of monkey patching to modify the `method_invocation` and `visit_expr` functions.
2. Override `get_cond` and `visit_switch_node` methods in a way that ensures the new versions of `method_invocation` and `visit_expr` are called.

If `method_invocation` is a method inside `JSONWriter`, the procedure is much simpler - the developer only needs to override `method_invocation`.